Renames modules and classes
