--- src/phonemes/sign/sign-writing-symbol.ts
@@ -1,3 +1,0 @@
-export type SignWritingSymbol = {
-    symbol: string
-};
