+++ src/phonemes/sign/sign-writing.ts
@@ -1,0 +1,7 @@
+export type SignWritingSymbol = {
+    symbol: string
+};
+
+export enum SignWritingCategory {
+    HANDS = "Hands"
+};
